refactor(test): split and clean forms.test.ts into focused subfiles
The 1680-line forms.test.ts had accumulated several quality problems:

- Implementation-detail tests checking Field object properties (.validate,
  .required, .autocomplete, .maxlength) rather than observable behaviour
- Global state mutations (setFormSuccess, setFormError, detectIframeMode)
  cleaned up inline, meaning any failed assertion would leave state
  polluted for subsequent tests — a concurrency risk
- Exhaustive valid-value tests for every accepted contact field enum value
  (8 tests that all called expectValid with minor variations)
- 11 mergeEventFields tests, several of which just confirmed "I put X in
  and got X back" without testing any merging logic

Changes:
- Deleted test/lib/forms.test.ts (1680 lines)
- Created test/lib/forms/{rendering,validate-form,flash,saved-data,
  event-fields,ticket-fields}.test.ts (1057 lines total)
- Replaced inline state cleanup with afterEach in CsrfForm and iframe tests
- Replaced field-property assertions with actual validation and renderFields
  HTML output tests (e.g. autocomplete attributes verified via rendered HTML)
- Consolidated repetitive mergeEventFields and contact-field-enum tests
- Added validateDate coverage for month 13 (already caught by Date NaN check)
- All 215 tests still pass
