--- test/lib/forms.test.ts
@@ -1,1680 +1,0 @@
-import { expect } from "@std/expect";
-import { afterEach, beforeAll, describe, it as test } from "@std/testing/bdd";
-import { stub } from "@std/testing/mock";
-import { getCurrentCsrfToken, signCsrfToken } from "#lib/csrf.ts";
-import { FormParams } from "#lib/form-data.ts";
-import {
-  CsrfForm,
-  clearSavedFormData,
-  type Field,
-  Flash,
-  renderError,
-  renderField,
-  renderFields,
-  renderSuccess,
-  setFormError,
-  setFormSuccess,
-  setSavedFormData,
-  validateForm,
-} from "#lib/forms.tsx";
-import { detectIframeMode } from "#lib/iframe.ts";
-import { MAX_TEXTAREA_LENGTH } from "#lib/limits.ts";
-import {
-  changePasswordFields,
-  eventFields,
-  fieldsApi,
-  getTicketFields,
-  groupCreateFields,
-  holidayFields,
-  joinFields,
-  loginFields,
-  mergeEventFields,
-  setupFields,
-  validateAddress,
-  validateBookableDays,
-  validateDate,
-  validatePhone,
-  validateSpecialInstructions,
-} from "#templates/fields.ts";
-import {
-  baseEventForm,
-  expectInvalid,
-  expectInvalidForm,
-  expectValid,
-  setupTestEncryptionKey,
-} from "#test-utils";
-
-beforeAll(async () => {
-  setupTestEncryptionKey();
-  await signCsrfToken();
-});
-
-/** Helper: build a Field definition with minimal boilerplate. */
-const field = (
-  overrides: Partial<Field> & { name: string; label: string },
-): Field => ({
-  type: "text",
-  ...overrides,
-});
-
-/** Helper: render a field with given overrides and optional value. */
-const rendered = (
-  overrides: Partial<Field> & { name: string; label: string },
-  value?: string,
-): string => renderField(field(overrides), value);
-
-/** Helper: build event form data with overrides. */
-const eventForm = (
-  overrides: Record<string, string> = {},
-): Record<string, string> => ({
-  ...baseEventForm,
-  ...overrides,
-});
-
-describe("forms", () => {
-  describe("renderField", () => {
-    test("renders text input with label", () => {
-      const html = rendered({ name: "username", label: "Username" });
-      expect(html).toContain("<label>");
-      expect(html).toContain("Username");
-      expect(html).toContain('type="text"');
-      expect(html).toContain('name="username"');
-      expect(html).toContain("</label>");
-    });
-
-    test("renders required attribute", () => {
-      const html = rendered({
-        name: "email",
-        label: "Email",
-        type: "email",
-        required: true,
-      });
-      expect(html).toContain("required");
-    });
-
-    test("renders placeholder", () => {
-      const html = rendered({
-        name: "name",
-        label: "Name",
-        placeholder: "Enter your name",
-      });
-      expect(html).toContain('placeholder="Enter your name"');
-    });
-
-    test("renders hint text", () => {
-      const html = rendered({
-        name: "password",
-        label: "Password",
-        type: "password",
-        hint: "Minimum 8 characters",
-      });
-      expect(html).toContain("Minimum 8 characters");
-      expect(html).toContain("<small");
-    });
-
-    test("renders hintHtml as raw HTML", () => {
-      const html = rendered({
-        name: "desc",
-        label: "Description",
-        hintHtml: '<a href="/guide">Help</a>',
-      });
-      expect(html).toContain('<a href="/guide">Help</a>');
-      expect(html).toContain("<small");
-    });
-
-    test("renders min attribute for number", () => {
-      const html = rendered({
-        name: "quantity",
-        label: "Quantity",
-        type: "number",
-        min: 1,
-      });
-      expect(html).toContain('min="1"');
-    });
-
-    test("renders pattern attribute", () => {
-      const html = rendered({
-        name: "code",
-        label: "Code",
-        pattern: "[A-Z]{3}",
-      });
-      expect(html).toContain('pattern="[A-Z]{3}"');
-    });
-
-    test("renders maxlength attribute", () => {
-      const html = rendered({
-        name: "desc",
-        label: "Description",
-        maxlength: 128,
-      });
-      expect(html).toContain('maxlength="128"');
-    });
-
-    test("renders autocomplete attribute when provided", () => {
-      const html = rendered({
-        name: "name",
-        label: "Name",
-        autocomplete: "name",
-      });
-      expect(html).toContain('autocomplete="name"');
-    });
-
-    test("omits autocomplete attribute when not provided", () => {
-      const html = rendered({ name: "name", label: "Name" });
-      expect(html).not.toContain("autocomplete");
-    });
-
-    test("renders textarea for textarea type", () => {
-      const html = rendered({
-        name: "description",
-        label: "Description",
-        type: "textarea",
-      });
-      expect(html).toContain("<textarea");
-      expect(html).not.toContain("<input");
-    });
-
-    test("renders value when provided", () => {
-      const html = rendered({ name: "name", label: "Name" }, "John Doe");
-      expect(html).toContain('value="John Doe"');
-    });
-
-    test("escapes HTML in value", () => {
-      const html = rendered(
-        { name: "name", label: "Name" },
-        '<script>alert("xss")</script>',
-      );
-      expect(html).toContain("&lt;script&gt;");
-      expect(html).not.toContain("<script>");
-    });
-
-    test("renders textarea with value", () => {
-      const html = rendered(
-        { name: "description", label: "Description", type: "textarea" },
-        "Some description",
-      );
-      expect(html).toContain(">Some description</textarea>");
-    });
-  });
-
-  describe("renderFields", () => {
-    test("renders multiple fields", () => {
-      const fields: Field[] = [
-        field({ name: "name", label: "Name", required: true }),
-        field({ name: "email", label: "Email", type: "email", required: true }),
-      ];
-      const html = renderFields(fields);
-      expect(html).toContain("Name");
-      expect(html).toContain("Email");
-      expect(html).toContain('name="name"');
-      expect(html).toContain('name="email"');
-    });
-
-    test("renders fields with values", () => {
-      const fields: Field[] = [
-        field({ name: "name", label: "Name" }),
-        field({ name: "count", label: "Count", type: "number" }),
-      ];
-      const values = { name: "Test", count: 42 };
-      const html = renderFields(fields, values);
-      expect(html).toContain('value="Test"');
-      expect(html).toContain('value="42"');
-    });
-
-    test("handles null values", () => {
-      const fields: Field[] = [
-        field({ name: "price", label: "Price", type: "number" }),
-      ];
-      const html = renderFields(fields, { price: null });
-      expect(html).not.toContain('value="null"');
-    });
-  });
-
-  describe("validateForm", () => {
-    const requiredName: Field[] = [
-      field({ name: "name", label: "Name", required: true }),
-    ];
-
-    test("validates required fields", () => {
-      expectInvalid("Name is required")(requiredName, { name: "" });
-    });
-
-    test("validates required field with whitespace only", () => {
-      expectInvalidForm(requiredName, { name: "   " });
-    });
-
-    test("passes validation when required field has value", () => {
-      const values = expectValid(requiredName, { name: "John" });
-      expect(values.name).toBe("John");
-    });
-
-    test("parses number fields", () => {
-      const fields: Field[] = [
-        field({
-          name: "quantity",
-          label: "Quantity",
-          type: "number",
-          required: true,
-        }),
-      ];
-      const values = expectValid(fields, { quantity: "42" });
-      expect(values.quantity).toBe(42);
-    });
-
-    test("returns null for empty optional number", () => {
-      const fields: Field[] = [
-        field({ name: "price", label: "Price", type: "number" }),
-      ];
-      const values = expectValid(fields, { price: "" });
-      expect(values.price).toBeNull();
-    });
-
-    test("returns empty string for empty optional text", () => {
-      const fields: Field[] = [field({ name: "note", label: "Note" })];
-      const values = expectValid(fields, { note: "" });
-      expect(values.note).toBe("");
-    });
-
-    test("runs custom validate function", () => {
-      const fields: Field[] = [
-        field({
-          name: "code",
-          label: "Code",
-          required: true,
-          validate: (v) =>
-            v.length !== 3 ? "Code must be 3 characters" : null,
-        }),
-      ];
-      expectInvalid("Code must be 3 characters")(fields, { code: "AB" });
-    });
-
-    test("skips custom validate for empty optional field", () => {
-      const fields: Field[] = [
-        field({
-          name: "code",
-          label: "Code",
-          validate: (v) =>
-            v.length !== 3 ? "Code must be 3 characters" : null,
-        }),
-      ];
-      expectValid(fields, { code: "" });
-    });
-
-    test("trims values", () => {
-      const values = expectValid(requiredName, { name: "  John  " });
-      expect(values.name).toBe("John");
-    });
-
-    test("collects checkbox-group values from multiple form entries", () => {
-      const fields: Field[] = [
-        field({ name: "days", label: "Days", type: "checkbox-group" }),
-      ];
-      const form = new FormParams();
-      form.append("days", "Monday");
-      form.append("days", "Wednesday");
-      const result = validateForm(form, fields);
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.days).toBe("Monday,Wednesday");
-      }
-    });
-
-    test("handles checkbox-group with single comma-separated value", () => {
-      const fields: Field[] = [
-        field({ name: "days", label: "Days", type: "checkbox-group" }),
-      ];
-      const result = validateForm(
-        new FormParams({ days: "Monday,Friday" }),
-        fields,
-      );
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.days).toBe("Monday,Friday");
-      }
-    });
-
-    test("returns empty string for empty checkbox-group", () => {
-      const fields: Field[] = [
-        field({ name: "days", label: "Days", type: "checkbox-group" }),
-      ];
-      const result = validateForm(new FormParams(), fields);
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.days).toBe("");
-      }
-    });
-
-    test("skips file fields in validation and returns null value", () => {
-      const fields: Field[] = [
-        field({ name: "image", label: "Image", type: "file" }),
-      ];
-      const result = validateForm(new FormParams(), fields);
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.image).toBeNull();
-      }
-    });
-  });
-
-  describe("renderError", () => {
-    test("returns empty string when no error", () => {
-      expect(renderError()).toBe("");
-      expect(renderError(undefined)).toBe("");
-    });
-
-    test("renders error message", () => {
-      const html = renderError("Something went wrong");
-      expect(html).toContain("Something went wrong");
-      expect(html).toContain('class="error"');
-      expect(html).toContain('role="alert"');
-    });
-
-    test("escapes HTML in error message", () => {
-      const html = renderError("<script>alert(1)</script>");
-      expect(html).toContain("&lt;script&gt;");
-      expect(html).not.toContain("<script>");
-    });
-  });
-
-  describe("Flash", () => {
-    test("renders nothing when no error or success", () => {
-      const html = String(Flash({}));
-      expect(html).not.toContain('class="error"');
-      expect(html).not.toContain('class="success"');
-    });
-
-    test("renders error message with role alert", () => {
-      const html = String(Flash({ error: "Bad input" }));
-      expect(html).toContain("Bad input");
-      expect(html).toContain('class="error"');
-      expect(html).toContain('role="alert"');
-      expect(html).not.toContain('class="success"');
-    });
-
-    test("renders success message with role alert", () => {
-      const html = String(Flash({ success: "Saved" }));
-      expect(html).toContain("Saved");
-      expect(html).toContain('class="success"');
-      expect(html).toContain('role="alert"');
-      expect(html).not.toContain('class="error"');
-    });
-
-    test("renders both error and success", () => {
-      const html = String(Flash({ error: "Oops", success: "Done" }));
-      expect(html).toContain("Oops");
-      expect(html).toContain("Done");
-      expect(html).toContain('class="error"');
-      expect(html).toContain('class="success"');
-    });
-
-    test("escapes HTML in messages", () => {
-      const html = String(Flash({ error: "<script>xss</script>" }));
-      expect(html).toContain("&lt;script&gt;");
-      expect(html).not.toContain("<script>xss");
-    });
-  });
-
-  describe("eventFields validation", () => {
-    test("validates thank_you_url rejects javascript: protocol", () => {
-      expectInvalid("URL must use https://")(
-        eventFields,
-        eventForm({ thank_you_url: "javascript:alert(1)" }),
-      );
-    });
-
-    test("validates thank_you_url rejects http: protocol", () => {
-      expectInvalid("URL must use https://")(
-        eventFields,
-        eventForm({ thank_you_url: "http://example.com/thank-you" }),
-      );
-    });
-
-    test("validates thank_you_url rejects invalid URL", () => {
-      expectInvalid("Invalid URL format")(
-        eventFields,
-        eventForm({ thank_you_url: "not-a-valid-url" }),
-      );
-    });
-
-    test("validates thank_you_url accepts relative URLs", () => {
-      expectValid(eventFields, eventForm({ thank_you_url: "/thank-you" }));
-    });
-
-    test("validates unit_price rejects negative values", () => {
-      expectInvalid("Price must be 0 or greater")(
-        eventFields,
-        eventForm({ unit_price: "-100" }),
-      );
-    });
-
-    test("validates max_price accepts valid values", () => {
-      expectValid(eventFields, eventForm({ max_price: "100.00" }));
-    });
-
-    test("validates max_price rejects negative values", () => {
-      expectInvalid("Price must be 0 or greater")(
-        eventFields,
-        eventForm({ max_price: "-50" }),
-      );
-    });
-
-    test("validates max_price accepts empty value", () => {
-      expectValid(eventFields, eventForm({ max_price: "" }));
-    });
-
-    test("validates name is required", () => {
-      const { name: _, ...formWithoutName } = baseEventForm;
-      expectInvalid("Event Name is required")(eventFields, formWithoutName);
-    });
-
-    test("validates description rejects values exceeding max length", () => {
-      const longDescription = "a".repeat(MAX_TEXTAREA_LENGTH + 1);
-      expectInvalid(
-        `Description must be ${MAX_TEXTAREA_LENGTH} characters or fewer`,
-      )(eventFields, eventForm({ description: longDescription }));
-    });
-
-    test("validates description accepts values within max length", () => {
-      expectValid(
-        eventFields,
-        eventForm({ description: "a".repeat(MAX_TEXTAREA_LENGTH) }),
-      );
-    });
-
-    test("validates description accepts empty value", () => {
-      expectValid(eventFields, eventForm({ description: "" }));
-    });
-
-    test("description field has maxlength matching MAX_TEXTAREA_LENGTH", () => {
-      const descField = eventFields.find((f) => f.name === "description");
-      expect(descField).toBeDefined();
-      expect(descField!.maxlength).toBe(MAX_TEXTAREA_LENGTH);
-    });
-  });
-
-  describe("groupCreateFields validation", () => {
-    test("rejects terms_and_conditions exceeding MAX_TEXTAREA_LENGTH", () => {
-      const termsField = groupCreateFields.find(
-        (f) => f.name === "terms_and_conditions",
-      )!;
-      const error = termsField.validate?.("x".repeat(MAX_TEXTAREA_LENGTH + 1));
-      expect(error).toContain(`${MAX_TEXTAREA_LENGTH} characters or fewer`);
-    });
-
-    test("accepts terms_and_conditions within MAX_TEXTAREA_LENGTH", () => {
-      const termsField = groupCreateFields.find(
-        (f) => f.name === "terms_and_conditions",
-      )!;
-      const error = termsField.validate?.("x".repeat(MAX_TEXTAREA_LENGTH));
-      expect(error).toBeNull();
-    });
-  });
-
-  describe("ticketFields validation", () => {
-    test("validates email format", () => {
-      expectInvalid("Please enter a valid email address")(
-        getTicketFields("email", false),
-        { name: "John Doe", email: "not-an-email" },
-      );
-    });
-
-    test("accepts valid email", () => {
-      expectValid(getTicketFields("email", false), {
-        name: "John Doe",
-        email: "john@example.com",
-      });
-    });
-  });
-
-  describe("renderField select type", () => {
-    const colorSelect: Field = {
-      name: "color",
-      label: "Color",
-      type: "select",
-      options: [
-        { value: "red", label: "Red" },
-        { value: "blue", label: "Blue" },
-      ],
-    };
-
-    test("renders select element with options", () => {
-      const html = renderField(colorSelect);
-      expect(html).toContain("<select");
-      expect(html).toContain('name="color"');
-      expect(html).toContain('value="red"');
-      expect(html).toContain(">Red</option>");
-      expect(html).toContain('value="blue"');
-      expect(html).toContain(">Blue</option>");
-    });
-
-    test("renders select with selected value", () => {
-      const html = renderField(colorSelect, "blue");
-      expect(html).toContain('value="blue" selected');
-      expect(html).not.toContain('value="red" selected');
-    });
-
-    test("renders select with hint", () => {
-      const selectWithHint: Field = {
-        name: "priority",
-        label: "Priority Level",
-        type: "select",
-        hint: "Choose the priority for this item",
-        options: [
-          { value: "low", label: "Low" },
-          { value: "medium", label: "Medium" },
-          { value: "high", label: "High" },
-        ],
-      };
-      const html = renderField(selectWithHint);
-      expect(html).toContain("Choose the priority for this item");
-    });
-  });
-
-  describe("validatePhone", () => {
-    test("accepts valid phone with country code", () => {
-      expect(validatePhone("+1 234 567 8900")).toBeNull();
-    });
-
-    test("accepts valid phone with parentheses", () => {
-      expect(validatePhone("+1 (555) 123-4567")).toBeNull();
-    });
-
-    test("accepts valid phone with hyphens", () => {
-      expect(validatePhone("+44-20-1234-5678")).toBeNull();
-    });
-
-    test("accepts plain digit phone", () => {
-      expect(validatePhone("1234567890")).toBeNull();
-    });
-
-    test("rejects phone too short", () => {
-      expect(validatePhone("123")).not.toBeNull();
-    });
-
-    test("rejects phone with letters", () => {
-      expect(validatePhone("abc1234567")).not.toBeNull();
-    });
-
-    test("rejects empty string", () => {
-      expect(validatePhone("")).not.toBeNull();
-    });
-  });
-
-  describe("validateAddress", () => {
-    test("accepts short address", () => {
-      expect(validateAddress("123 Main St")).toBeNull();
-    });
-
-    test("accepts address at max length (250)", () => {
-      expect(validateAddress("a".repeat(250))).toBeNull();
-    });
-
-    test("rejects address exceeding max length", () => {
-      expect(validateAddress("a".repeat(251))).toBe(
-        "Address must be 250 characters or fewer",
-      );
-    });
-
-    test("accepts multi-line address within limit", () => {
-      expect(
-        validateAddress("123 Main St\nApt 4\nSpringfield, IL 62701"),
-      ).toBeNull();
-    });
-  });
-
-  describe("validateSpecialInstructions", () => {
-    test("accepts short special instructions", () => {
-      expect(validateSpecialInstructions("No nuts please")).toBeNull();
-    });
-
-    test("accepts special instructions at max length (250)", () => {
-      expect(validateSpecialInstructions("a".repeat(250))).toBeNull();
-    });
-
-    test("rejects special instructions exceeding max length", () => {
-      expect(validateSpecialInstructions("a".repeat(251))).toBe(
-        "Special instructions must be 250 characters or fewer",
-      );
-    });
-
-    test("accepts multi-line special instructions within limit", () => {
-      expect(validateSpecialInstructions("Line 1\nLine 2\nLine 3")).toBeNull();
-    });
-  });
-
-  describe("getTicketFields", () => {
-    test("returns name and email fields for email setting", () => {
-      const fields = getTicketFields("email", false);
-      expect(fields.length).toBe(2);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("email");
-    });
-
-    test("returns name and phone fields for phone setting", () => {
-      const fields = getTicketFields("phone", false);
-      expect(fields.length).toBe(2);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("phone");
-    });
-
-    test("returns name, email, and phone fields for email,phone setting", () => {
-      const fields = getTicketFields("email,phone", false);
-      expect(fields.length).toBe(3);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("email");
-      expect(fields[2]!.name).toBe("phone");
-    });
-
-    test("phone field has validation", () => {
-      const phoneField = getTicketFields("phone", false)[1]!;
-      expect(phoneField.validate).toBeDefined();
-      expect(phoneField.required).toBe(true);
-    });
-
-    test("email field has validation", () => {
-      const emailField = getTicketFields("email", false)[1]!;
-      expect(emailField.validate).toBeDefined();
-      expect(emailField.required).toBe(true);
-    });
-
-    test("returns name and address fields for address setting", () => {
-      const fields = getTicketFields("address", false);
-      expect(fields.length).toBe(2);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("address");
-    });
-
-    test("returns all four contact fields for email,phone,address setting", () => {
-      const fields = getTicketFields("email,phone,address", false);
-      expect(fields.length).toBe(4);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("email");
-      expect(fields[2]!.name).toBe("phone");
-      expect(fields[3]!.name).toBe("address");
-    });
-
-    test("address field has validation", () => {
-      const addrField = getTicketFields("address", false)[1]!;
-      expect(addrField.validate).toBeDefined();
-      expect(addrField.required).toBe(true);
-      expect(addrField.type).toBe("textarea");
-    });
-
-    test("returns name and special_instructions fields for special_instructions setting", () => {
-      const fields = getTicketFields("special_instructions", false);
-      expect(fields.length).toBe(2);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("special_instructions");
-    });
-
-    test("special_instructions field has validation", () => {
-      const field = getTicketFields("special_instructions", false)[1]!;
-      expect(field.validate).toBeDefined();
-      expect(field.required).toBe(true);
-      expect(field.type).toBe("textarea");
-    });
-
-    test("returns all five contact fields for email,phone,address,special_instructions setting", () => {
-      const fields = getTicketFields(
-        "email,phone,address,special_instructions",
-        false,
-      );
-      expect(fields.length).toBe(5);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("email");
-      expect(fields[2]!.name).toBe("phone");
-      expect(fields[3]!.name).toBe("address");
-      expect(fields[4]!.name).toBe("special_instructions");
-    });
-
-    test("ignores unknown field names in comma-separated setting", () => {
-      const fields = getTicketFields("email,bogus,phone", false);
-      expect(fields.length).toBe(3);
-      expect(fields[0]!.name).toBe("name");
-      expect(fields[1]!.name).toBe("email");
-      expect(fields[2]!.name).toBe("phone");
-    });
-
-    test("returns only name for empty fields setting", () => {
-      const fields = getTicketFields("", false);
-      expect(fields.length).toBe(1);
-      expect(fields[0]!.name).toBe("name");
-    });
-
-    test("name field has autocomplete=name", () => {
-      const fields = getTicketFields("email", false);
-      expect(fields[0]!.autocomplete).toBe("name");
-    });
-
-    test("email field has autocomplete=email", () => {
-      const fields = getTicketFields("email", false);
-      expect(fields[1]!.autocomplete).toBe("email");
-    });
-
-    test("phone field has autocomplete=tel", () => {
-      const fields = getTicketFields("phone", false);
-      expect(fields[1]!.autocomplete).toBe("tel");
-    });
-
-    test("address field has autocomplete=street-address", () => {
-      const fields = getTicketFields("address", false);
-      expect(fields[1]!.autocomplete).toBe("street-address");
-    });
-
-    test("includes email when Square is active and event is paid", () => {
-      const s = stub(fieldsApi, "getSettingCached", () => "square");
-      try {
-        const fields = getTicketFields("phone", true);
-        expect(fields.length).toBe(3);
-        expect(fields[0]!.name).toBe("name");
-        expect(fields[1]!.name).toBe("email");
-        expect(fields[2]!.name).toBe("phone");
-      } finally {
-        s.restore();
-      }
-    });
-
-    test("does not add email with Square when event is free", () => {
-      const s = stub(fieldsApi, "getSettingCached", () => "square");
-      try {
-        const fields = getTicketFields("phone", false);
-        expect(fields.length).toBe(2);
-        expect(fields[0]!.name).toBe("name");
-        expect(fields[1]!.name).toBe("phone");
-      } finally {
-        s.restore();
-      }
-    });
-
-    test("does not duplicate email with Square when already present", () => {
-      const s = stub(fieldsApi, "getSettingCached", () => "square");
-      try {
-        const fields = getTicketFields("email,phone", true);
-        expect(fields.length).toBe(3);
-        expect(fields[0]!.name).toBe("name");
-        expect(fields[1]!.name).toBe("email");
-        expect(fields[2]!.name).toBe("phone");
-      } finally {
-        s.restore();
-      }
-    });
-
-    test("adds email with Square even for empty fields when paid", () => {
-      const s = stub(fieldsApi, "getSettingCached", () => "square");
-      try {
-        const fields = getTicketFields("", true);
-        expect(fields.length).toBe(2);
-        expect(fields[0]!.name).toBe("name");
-        expect(fields[1]!.name).toBe("email");
-      } finally {
-        s.restore();
-      }
-    });
-  });
-
-  describe("mergeEventFields", () => {
-    test("returns empty string for empty array", () => {
-      expect(mergeEventFields([])).toBe("");
-    });
-
-    test("returns empty string when all events have empty fields", () => {
-      expect(mergeEventFields(["", ""])).toBe("");
-    });
-
-    test("returns email when all events use email", () => {
-      expect(mergeEventFields(["email", "email", "email"])).toBe("email");
-    });
-
-    test("returns phone when all events use phone", () => {
-      expect(mergeEventFields(["phone", "phone"])).toBe("phone");
-    });
-
-    test("returns email,phone when all events use email,phone", () => {
-      expect(mergeEventFields(["email,phone", "email,phone"])).toBe(
-        "email,phone",
-      );
-    });
-
-    test("returns email,phone when events differ (email + phone)", () => {
-      expect(mergeEventFields(["email", "phone"])).toBe("email,phone");
-    });
-
-    test("returns email,phone when events differ (email + email,phone)", () => {
-      expect(mergeEventFields(["email", "email,phone"])).toBe("email,phone");
-    });
-
-    test("returns email,phone when events differ (phone + email,phone)", () => {
-      expect(mergeEventFields(["phone", "email,phone"])).toBe("email,phone");
-    });
-
-    test("returns setting for single event", () => {
-      expect(mergeEventFields(["phone"])).toBe("phone");
-    });
-
-    test("returns email,phone,address for all fields combined", () => {
-      expect(mergeEventFields(["email", "phone,address"])).toBe(
-        "email,phone,address",
-      );
-    });
-
-    test("sorts output in canonical CONTACT_FIELDS order", () => {
-      expect(mergeEventFields(["address", "email"])).toBe("email,address");
-    });
-
-    test("includes special_instructions when merged", () => {
-      expect(mergeEventFields(["email,special_instructions", "phone"])).toBe(
-        "email,phone,special_instructions",
-      );
-    });
-  });
-
-  describe("eventFields Contact Fields validation", () => {
-    test("validates fields select rejects invalid value", () => {
-      expectInvalid("Invalid contact field: invalid")(
-        eventFields,
-        eventForm({ fields: "invalid" }),
-      );
-    });
-
-    test("validates fields select accepts email", () => {
-      expectValid(eventFields, eventForm({ fields: "email" }));
-    });
-
-    test("validates fields select accepts phone", () => {
-      expectValid(eventFields, eventForm({ fields: "phone" }));
-    });
-
-    test("validates fields select accepts email,phone", () => {
-      expectValid(eventFields, eventForm({ fields: "email,phone" }));
-    });
-
-    test("validates fields select accepts address", () => {
-      expectValid(eventFields, eventForm({ fields: "address" }));
-    });
-
-    test("validates fields select accepts email,phone,address", () => {
-      expectValid(eventFields, eventForm({ fields: "email,phone,address" }));
-    });
-
-    test("validates fields select accepts special_instructions", () => {
-      expectValid(eventFields, eventForm({ fields: "special_instructions" }));
-    });
-
-    test("validates fields select accepts email,phone,address,special_instructions", () => {
-      expectValid(
-        eventFields,
-        eventForm({ fields: "email,phone,address,special_instructions" }),
-      );
-    });
-  });
-
-  describe("eventFields Event Type validation", () => {
-    test("validates event_type accepts standard", () => {
-      expectValid(eventFields, eventForm({ event_type: "standard" }));
-    });
-
-    test("validates event_type accepts daily", () => {
-      expectValid(eventFields, eventForm({ event_type: "daily" }));
-    });
-
-    test("validates event_type rejects invalid value", () => {
-      expectInvalid("Event Type must be standard or daily")(
-        eventFields,
-        eventForm({ event_type: "weekly" }),
-      );
-    });
-
-    test("accepts empty event_type (optional)", () => {
-      expectValid(eventFields, eventForm());
-    });
-  });
-
-  describe("eventFields Bookable Days validation", () => {
-    test("validates bookable_days accepts valid days", () => {
-      expectValid(
-        eventFields,
-        eventForm({ bookable_days: "Monday,Wednesday,Friday" }),
-      );
-    });
-
-    test("validates bookable_days accepts all days", () => {
-      expectValid(
-        eventFields,
-        eventForm({
-          bookable_days:
-            "Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday",
-        }),
-      );
-    });
-
-    test("validates bookable_days rejects invalid day name", () => {
-      expectInvalid(
-        "Invalid day: Funday. Use: Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday",
-      )(eventFields, eventForm({ bookable_days: "Monday,Funday" }));
-    });
-
-    test("validates bookable_days rejects empty after trimming", () => {
-      expectInvalid("At least one day is required")(
-        eventFields,
-        eventForm({ bookable_days: "," }),
-      );
-    });
-
-    test("accepts empty bookable_days (optional)", () => {
-      expectValid(eventFields, eventForm());
-    });
-  });
-
-  describe("phone ticket fields validation", () => {
-    test("validates phone is required for phone-only events", () => {
-      expectInvalid("Your Phone Number is required")(
-        getTicketFields("phone", false),
-        {
-          name: "John Doe",
-          phone: "",
-        },
-      );
-    });
-
-    test("validates phone format for phone-only events", () => {
-      expectInvalid("Please enter a valid phone number")(
-        getTicketFields("phone", false),
-        { name: "John Doe", phone: "abc" },
-      );
-    });
-
-    test("accepts valid phone for phone-only events", () => {
-      expectValid(getTicketFields("phone", false), {
-        name: "John Doe",
-        phone: "+1 555 123 4567",
-      });
-    });
-
-    test("validates both email and phone for email,phone setting", () => {
-      expectValid(getTicketFields("email,phone", false), {
-        name: "John Doe",
-        email: "john@example.com",
-        phone: "+1 555 123 4567",
-      });
-    });
-
-    test("rejects missing phone for email,phone setting", () => {
-      expectInvalidForm(getTicketFields("email,phone", false), {
-        name: "John Doe",
-        email: "john@example.com",
-        phone: "",
-      });
-    });
-
-    test("rejects missing email for email,phone setting", () => {
-      expectInvalidForm(getTicketFields("email,phone", false), {
-        name: "John Doe",
-        email: "",
-        phone: "+1 555 123 4567",
-      });
-    });
-  });
-
-  describe("holidayFields validation", () => {
-    const holidayForm = (
-      overrides: Record<string, string> = {},
-    ): Record<string, string> => ({
-      name: "Bank Holiday",
-      start_date: "2026-12-25",
-      end_date: "2026-12-25",
-      ...overrides,
-    });
-
-    test("validates required name", () => {
-      expectInvalid("Holiday Name is required")(
-        holidayFields,
-        holidayForm({ name: "" }),
-      );
-    });
-
-    test("validates required start_date", () => {
-      expectInvalid("Start Date is required")(
-        holidayFields,
-        holidayForm({ start_date: "" }),
-      );
-    });
-
-    test("validates required end_date", () => {
-      expectInvalid("End Date is required")(
-        holidayFields,
-        holidayForm({ end_date: "" }),
-      );
-    });
-
-    test("validates start_date format", () => {
-      expectInvalid("Please enter a valid date (YYYY-MM-DD)")(
-        holidayFields,
-        holidayForm({ start_date: "25-12-2026" }),
-      );
-    });
-
-    test("validates end_date format", () => {
-      expectInvalid("Please enter a valid date (YYYY-MM-DD)")(
-        holidayFields,
-        holidayForm({ end_date: "not-a-date" }),
-      );
-    });
-
-    test("accepts valid holiday form", () => {
-      const values = expectValid(holidayFields, holidayForm());
-      expect(values.name).toBe("Bank Holiday");
-      expect(values.start_date).toBe("2026-12-25");
-      expect(values.end_date).toBe("2026-12-25");
-    });
-
-    test("accepts multi-day holiday", () => {
-      expectValid(
-        holidayFields,
-        holidayForm({ start_date: "2026-12-24", end_date: "2026-12-26" }),
-      );
-    });
-  });
-
-  describe("validateDate", () => {
-    test("accepts valid date", () => {
-      expect(validateDate("2026-12-25")).toBeNull();
-    });
-
-    test("accepts leap year date", () => {
-      expect(validateDate("2028-02-29")).toBeNull();
-    });
-
-    test("rejects wrong format", () => {
-      expect(validateDate("12/25/2026")).toBe(
-        "Please enter a valid date (YYYY-MM-DD)",
-      );
-    });
-
-    test("rejects partial date", () => {
-      expect(validateDate("2026-12")).toBe(
-        "Please enter a valid date (YYYY-MM-DD)",
-      );
-    });
-
-    test("rejects text", () => {
-      expect(validateDate("not-a-date")).toBe(
-        "Please enter a valid date (YYYY-MM-DD)",
-      );
-    });
-
-    test("rejects empty string", () => {
-      expect(validateDate("")).toBe("Please enter a valid date (YYYY-MM-DD)");
-    });
-
-    test("rejects valid format but invalid date (month 00)", () => {
-      expect(validateDate("2026-00-01")).toBe("Please enter a valid date");
-    });
-  });
-
-  describe("renderField date type", () => {
-    test("renders date input", () => {
-      const html = rendered({
-        name: "start_date",
-        label: "Start Date",
-        type: "date",
-      });
-      expect(html).toContain('type="date"');
-      expect(html).toContain('name="start_date"');
-    });
-
-    test("renders date input with value", () => {
-      const html = rendered(
-        { name: "start_date", label: "Start Date", type: "date" },
-        "2026-12-25",
-      );
-      expect(html).toContain('value="2026-12-25"');
-    });
-  });
-
-  describe("renderField datetime type", () => {
-    test("renders split date and time inputs", () => {
-      const html = rendered({
-        name: "closes_at",
-        label: "Closes At",
-        type: "datetime",
-      });
-      expect(html).toContain('type="date"');
-      expect(html).toContain('name="closes_at_date"');
-      expect(html).toContain('placeholder="Date"');
-      expect(html).toContain('type="time"');
-      expect(html).toContain('name="closes_at_time"');
-      expect(html).toContain('placeholder="Time"');
-    });
-
-    test("renders split inputs with value", () => {
-      const html = rendered(
-        { name: "closes_at", label: "Closes At", type: "datetime" },
-        "2099-06-15T14:30",
-      );
-      expect(html).toContain('value="2099-06-15"');
-      expect(html).toContain('value="14:30"');
-    });
-
-    test("renders split inputs without value", () => {
-      const html = rendered(
-        { name: "closes_at", label: "Closes At", type: "datetime" },
-        "",
-      );
-      expect(html).not.toContain("value=");
-    });
-  });
-
-  describe("validateForm datetime type", () => {
-    const datetimeField: Field[] = [
-      field({ name: "closes_at", label: "Closes At", type: "datetime" }),
-    ];
-
-    test("combines date and time into datetime string", () => {
-      const result = validateForm(
-        new FormParams({
-          closes_at_date: "2099-06-15",
-          closes_at_time: "14:30",
-        }),
-        datetimeField,
-      );
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.closes_at).toBe("2099-06-15T14:30");
-      }
-    });
-
-    test("returns null when both date and time are empty", () => {
-      const result = validateForm(
-        new FormParams({ closes_at_date: "", closes_at_time: "" }),
-        datetimeField,
-      );
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.closes_at).toBeNull();
-      }
-    });
-
-    test("rejects empty required datetime", () => {
-      const requiredDatetime: Field[] = [
-        field({
-          name: "closes_at",
-          label: "Closes At",
-          type: "datetime",
-          required: true,
-        }),
-      ];
-      const result = validateForm(
-        new FormParams({ closes_at_date: "", closes_at_time: "" }),
-        requiredDatetime,
-      );
-      expect(result.valid).toBe(false);
-      if (!result.valid) {
-        expect(result.error).toBe("Closes At is required");
-      }
-    });
-
-    test("defaults time to 00:00 when only date is provided", () => {
-      const result = validateForm(
-        new FormParams({
-          closes_at_date: "2099-06-15",
-          closes_at_time: "",
-        }),
-        datetimeField,
-      );
-      expect(result.valid).toBe(true);
-      if (result.valid) {
-        expect(result.values.closes_at).toBe("2099-06-15T00:00");
-      }
-    });
-
-    test("rejects when only time is provided", () => {
-      const result = validateForm(
-        new FormParams({ closes_at_date: "", closes_at_time: "14:30" }),
-        datetimeField,
-      );
-      expect(result.valid).toBe(false);
-      if (!result.valid) {
-        expect(result.error).toBe(
-          "Please enter a date when providing a time, or leave both blank",
-        );
-      }
-    });
-  });
-
-  describe("renderField file type", () => {
-    test("renders file input with accept attribute", () => {
-      const html = rendered({
-        name: "image",
-        label: "Upload Image",
-        type: "file",
-        accept: "image/jpeg,image/png",
-      });
-      expect(html).toContain('type="file"');
-      expect(html).toContain('name="image"');
-      expect(html).toContain('accept="image/jpeg,image/png"');
-      expect(html).toContain("Upload Image");
-    });
-  });
-
-  describe("renderField checkbox-group type", () => {
-    test("renders checkbox group with options", () => {
-      const html = rendered({
-        name: "days",
-        label: "Days",
-        type: "checkbox-group",
-        options: [
-          { value: "Monday", label: "Monday" },
-          { value: "Tuesday", label: "Tuesday" },
-        ],
-      });
-      expect(html).toContain('type="checkbox"');
-      expect(html).toContain('name="days"');
-      expect(html).toContain('value="Monday"');
-      expect(html).toContain('value="Tuesday"');
-    });
-
-    test("renders checkbox group with pre-selected values", () => {
-      const html = rendered(
-        {
-          name: "days",
-          label: "Days",
-          type: "checkbox-group",
-          options: [
-            { value: "Monday", label: "Monday" },
-            { value: "Tuesday", label: "Tuesday" },
-            { value: "Wednesday", label: "Wednesday" },
-          ],
-        },
-        "Monday,Wednesday",
-      );
-      expect(html).toContain('value="Monday" checked');
-      expect(html).toContain('value="Wednesday" checked');
-      expect(html).not.toContain('value="Tuesday" checked');
-    });
-
-    test("renders empty checkbox group when no values selected", () => {
-      const html = rendered({
-        name: "days",
-        label: "Days",
-        type: "checkbox-group",
-        options: [{ value: "Monday", label: "Monday" }],
-      });
-      expect(html).not.toContain("checked");
-    });
-  });
-
-  describe("field autocomplete attributes", () => {
-    test("login username field has autocomplete=username", () => {
-      expect(loginFields[0]!.autocomplete).toBe("username");
-    });
-
-    test("login password field has autocomplete=current-password", () => {
-      expect(loginFields[1]!.autocomplete).toBe("current-password");
-    });
-
-    test("setup password fields have autocomplete=new-password", () => {
-      const pwField = setupFields.find((f) => f.name === "admin_password");
-      const confirmField = setupFields.find(
-        (f) => f.name === "admin_password_confirm",
-      );
-      expect(pwField!.autocomplete).toBe("new-password");
-      expect(confirmField!.autocomplete).toBe("new-password");
-    });
-
-    test("change password current field has autocomplete=current-password", () => {
-      expect(changePasswordFields[0]!.autocomplete).toBe("current-password");
-    });
-
-    test("change password new fields have autocomplete=new-password", () => {
-      expect(changePasswordFields[1]!.autocomplete).toBe("new-password");
-      expect(changePasswordFields[2]!.autocomplete).toBe("new-password");
-    });
-
-    test("join password fields have autocomplete=new-password", () => {
-      expect(joinFields[0]!.autocomplete).toBe("new-password");
-      expect(joinFields[1]!.autocomplete).toBe("new-password");
-    });
-  });
-
-  describe("validateBookableDays", () => {
-    test("accepts single valid day", () => {
-      expect(validateBookableDays("Monday")).toBeNull();
-    });
-
-    test("accepts multiple valid days", () => {
-      expect(validateBookableDays("Monday,Wednesday,Friday")).toBeNull();
-    });
-
-    test("accepts all days of the week", () => {
-      expect(
-        validateBookableDays(
-          "Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday",
-        ),
-      ).toBeNull();
-    });
-
-    test("trims whitespace around day names", () => {
-      expect(validateBookableDays(" Monday , Friday ")).toBeNull();
-    });
-
-    test("rejects invalid day name", () => {
-      expect(validateBookableDays("Monday,Funday")).toContain(
-        "Invalid day: Funday",
-      );
-    });
-
-    test("rejects empty string after splitting", () => {
-      expect(validateBookableDays(",")).toBe("At least one day is required");
-    });
-
-    test("rejects completely empty value", () => {
-      expect(validateBookableDays("  ")).toBe("At least one day is required");
-    });
-  });
-
-  describe("CsrfForm", () => {
-    test("renders form with POST method and action", () => {
-      const html = String(CsrfForm({ action: "/submit" }));
-      expect(html).toContain('<form method="POST" action="/submit"');
-      expect(html).toContain("</form>");
-    });
-
-    test("includes hidden csrf_token input with stored token", () => {
-      const html = String(CsrfForm({ action: "/submit" }));
-      expect(html).toContain(
-        `<input type="hidden" name="csrf_token" value="${getCurrentCsrfToken()}"`,
-      );
-    });
-
-    test("passes through class attribute", () => {
-      const html = String(CsrfForm({ action: "/submit", class: "inline" }));
-      expect(html).toContain('class="inline"');
-    });
-
-    test("passes through enctype for multipart forms", () => {
-      const html = String(
-        CsrfForm({ action: "/upload", enctype: "multipart/form-data" }),
-      );
-      expect(html).toContain('enctype="multipart/form-data"');
-    });
-
-    test("renders children inside the form", () => {
-      const html = String(
-        CsrfForm({ action: "/submit", children: "Submit here" }),
-      );
-      expect(html).toContain("Submit here");
-      expect(html).toContain("</form>");
-    });
-
-    test("renders id attribute when provided", () => {
-      const html = String(
-        CsrfForm({ action: "/submit", id: "settings-timezone" }),
-      );
-      expect(html).toContain('id="settings-timezone"');
-    });
-
-    test("does not render id attribute when not provided", () => {
-      const html = String(CsrfForm({ action: "/submit" }));
-      expect(html).not.toContain("id=");
-    });
-
-    test("renders success message when id matches stored success state", () => {
-      setFormSuccess("my-form", "Saved");
-      const html = String(CsrfForm({ action: "/submit", id: "my-form" }));
-      expect(html).toContain("Saved");
-      expect(html).toContain('class="success"');
-      expect(html).toContain('role="alert"');
-      setFormSuccess("", "");
-    });
-
-    test("does not render success message when id does not match", () => {
-      setFormSuccess("other-form", "Saved");
-      const html = String(CsrfForm({ action: "/submit", id: "my-form" }));
-      expect(html).not.toContain('class="success"');
-      setFormSuccess("", "");
-    });
-
-    test("does not render success message when no id", () => {
-      setFormSuccess("my-form", "Saved");
-      const html = String(CsrfForm({ action: "/submit" }));
-      expect(html).not.toContain('class="success"');
-      setFormSuccess("", "");
-    });
-
-    test("renders error message when id matches stored error state", () => {
-      setFormError("my-form", "Something went wrong");
-      const html = String(CsrfForm({ action: "/submit", id: "my-form" }));
-      expect(html).toContain("Something went wrong");
-      expect(html).toContain('class="error"');
-      expect(html).toContain('role="alert"');
-      setFormError("", "");
-    });
-
-    test("does not render error message when id does not match", () => {
-      setFormError("other-form", "Something went wrong");
-      const html = String(CsrfForm({ action: "/submit", id: "my-form" }));
-      expect(html).not.toContain('class="error"');
-      setFormError("", "");
-    });
-
-    test("does not render error message when no id", () => {
-      setFormError("my-form", "Something went wrong");
-      const html = String(CsrfForm({ action: "/submit" }));
-      expect(html).not.toContain('class="error"');
-      setFormError("", "");
-    });
-
-    test("appends ?iframe=true to action when iframe mode is active", () => {
-      detectIframeMode("https://example.com/?iframe=true");
-      const html = String(CsrfForm({ action: "/ticket/test" }));
-      expect(html).toContain('action="/ticket/test?iframe=true"');
-      detectIframeMode("https://example.com/");
-    });
-
-    test("does not append iframe param when iframe mode is inactive", () => {
-      detectIframeMode("https://example.com/");
-      const html = String(CsrfForm({ action: "/ticket/test" }));
-      expect(html).toContain('action="/ticket/test"');
-      expect(html).not.toContain("iframe");
-    });
-  });
-
-  describe("renderSuccess", () => {
-    test("returns empty string when no message", () => {
-      expect(renderSuccess()).toBe("");
-      expect(renderSuccess(undefined)).toBe("");
-    });
-
-    test("renders success message", () => {
-      const html = renderSuccess("Changes saved");
-      expect(html).toContain("Changes saved");
-      expect(html).toContain('class="success"');
-      expect(html).toContain('role="alert"');
-    });
-
-    test("escapes HTML in success message", () => {
-      const html = renderSuccess("<script>alert(1)</script>");
-      expect(html).toContain("&lt;script&gt;");
-      expect(html).not.toContain("<script>");
-    });
-  });
-
-  describe("saved form data preservation", () => {
-    const nameField = field({ name: "name", label: "Name" });
-    const emailField = field({ name: "email", label: "Email", type: "email" });
-
-    afterEach(() => {
-      clearSavedFormData();
-    });
-
-    test("renderFields restores saved values when no explicit values provided", () => {
-      setSavedFormData(new FormParams("name=Alice&email=alice%40test.com"));
-      const html = renderFields([nameField, emailField]);
-      expect(html).toContain('value="Alice"');
-      expect(html).toContain('value="alice@test.com"');
-    });
-
-    test("explicit values take precedence over saved values", () => {
-      setSavedFormData(new FormParams("name=SavedName"));
-      const html = renderFields([nameField], { name: "ExplicitName" });
-      expect(html).toContain('value="ExplicitName"');
-      expect(html).not.toContain("SavedName");
-    });
-
-    test("does not restore password fields", () => {
-      const pwField = field({
-        name: "password",
-        label: "Password",
-        type: "password",
-      });
-      setSavedFormData(new FormParams("password=secret123"));
-      const html = renderFields([pwField]);
-      expect(html).not.toContain("secret123");
-    });
-
-    test("does not restore file fields", () => {
-      const fileField = field({
-        name: "photo",
-        label: "Photo",
-        type: "file",
-      });
-      setSavedFormData(new FormParams("photo=hack.jpg"));
-      const html = renderFields([fileField]);
-      expect(html).not.toContain("hack.jpg");
-    });
-
-    test("restores textarea values", () => {
-      const textareaField = field({
-        name: "notes",
-        label: "Notes",
-        type: "textarea",
-      });
-      setSavedFormData(new FormParams("notes=My+notes+here"));
-      const html = renderFields([textareaField]);
-      expect(html).toContain("My notes here");
-    });
-
-    test("restores select values", () => {
-      const selectField = field({
-        name: "color",
-        label: "Color",
-        type: "select",
-        options: [
-          { value: "red", label: "Red" },
-          { value: "blue", label: "Blue" },
-        ],
-      });
-      setSavedFormData(new FormParams("color=blue"));
-      const html = renderFields([selectField]);
-      expect(html).toContain('value="blue" selected');
-    });
-
-    test("restores checkbox-group values", () => {
-      const checkboxField = field({
-        name: "tags",
-        label: "Tags",
-        type: "checkbox-group",
-        options: [
-          { value: "a", label: "A" },
-          { value: "b", label: "B" },
-          { value: "c", label: "C" },
-        ],
-      });
-      setSavedFormData(new FormParams("tags=a&tags=c"));
-      const html = renderFields([checkboxField]);
-      expect(html).toContain('value="a" checked');
-      expect(html).not.toContain('value="b" checked');
-      expect(html).toContain('value="c" checked');
-    });
-
-    test("restores datetime field values", () => {
-      const dtField = field({
-        name: "start",
-        label: "Start",
-        type: "datetime",
-      });
-      setSavedFormData(
-        new FormParams("start_date=2026-03-21&start_time=14%3A30"),
-      );
-      const html = renderFields([dtField]);
-      expect(html).toContain('value="2026-03-21"');
-      expect(html).toContain('value="14:30"');
-    });
-
-    test("restores datetime field with no date or time as empty", () => {
-      const dtField = field({
-        name: "start",
-        label: "Start",
-        type: "datetime",
-      });
-      setSavedFormData(new FormParams("other=value"));
-      const html = renderFields([dtField]);
-      expect(html).not.toContain('value="');
-    });
-
-    test("restores datetime field with date only", () => {
-      const dtField = field({
-        name: "start",
-        label: "Start",
-        type: "datetime",
-      });
-      setSavedFormData(new FormParams("start_date=2026-03-21"));
-      const html = renderFields([dtField]);
-      expect(html).toContain('value="2026-03-21"');
-      expect(html).toContain('value="00:00"');
-    });
-
-    test("clearSavedFormData removes saved data", () => {
-      setSavedFormData(new FormParams("name=Alice"));
-      clearSavedFormData();
-      const html = renderFields([nameField]);
-      expect(html).not.toContain("Alice");
-    });
-
-    test("returns empty values when no saved data exists", () => {
-      const html = renderFields([nameField]);
-      expect(html).not.toContain('value="');
-    });
-
-    test("escapes HTML in saved values", () => {
-      setSavedFormData(
-        new FormParams("name=%3Cscript%3Ealert(1)%3C%2Fscript%3E"),
-      );
-      const html = renderFields([nameField]);
-      expect(html).toContain("&lt;script&gt;");
-      expect(html).not.toContain("<script>alert");
-    });
-
-    test("excludes csrf_token from saved data restoration", () => {
-      setSavedFormData(new FormParams("csrf_token=old_token&name=Alice"));
-      const csrfField = field({ name: "csrf_token", label: "CSRF" });
-      const html = renderFields([nameField, csrfField]);
-      expect(html).toContain('value="Alice"');
-      // csrf_token is rendered via CsrfForm hidden input, not renderFields
-      // but even if a csrf_token field existed, it would just restore the old value
-      // which is harmless since CsrfForm overwrites it
-    });
-
-    test("defaultValue is used when no saved data and no explicit value", () => {
-      const fieldWithDefault = field({
-        name: "country",
-        label: "Country",
-        defaultValue: "US",
-      });
-      const html = renderFields([fieldWithDefault]);
-      expect(html).toContain('value="US"');
-    });
-
-    test("saved data takes precedence over defaultValue", () => {
-      const fieldWithDefault = field({
-        name: "country",
-        label: "Country",
-        defaultValue: "US",
-      });
-      setSavedFormData(new FormParams("country=GB"));
-      const html = renderFields([fieldWithDefault]);
-      expect(html).toContain('value="GB"');
-      expect(html).not.toContain('value="US"');
-    });
-  });
-});
