+++ test/lib/forms/validate-form.test.ts
@@ -1,0 +1,146 @@
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { type Field, validateForm } from "#lib/forms.tsx";
+import { FormParams } from "#lib/form-data.ts";
+
+const field = (
+  overrides: Partial<Field> & { name: string; label: string },
+): Field => ({ type: "text", ...overrides });
+
+const requiredName: Field[] = [field({ name: "name", label: "Name", required: true })];
+
+describe("validateForm", () => {
+  test("rejects empty required field", () => {
+    const result = validateForm(new FormParams({ name: "" }), requiredName);
+    expect(result.valid).toBe(false);
+    if (!result.valid) expect(result.error).toBe("Name is required");
+  });
+
+  test("rejects whitespace-only required field", () => {
+    expect(validateForm(new FormParams({ name: "   " }), requiredName).valid).toBe(false);
+  });
+
+  test("passes required field with value and trims it", () => {
+    const result = validateForm(new FormParams({ name: "  John  " }), requiredName);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.name).toBe("John");
+  });
+
+  test("parses number field to a numeric value", () => {
+    const fields: Field[] = [field({ name: "qty", label: "Qty", type: "number", required: true })];
+    const result = validateForm(new FormParams({ qty: "42" }), fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.qty).toBe(42);
+  });
+
+  test("returns null for empty optional number", () => {
+    const fields: Field[] = [field({ name: "price", label: "Price", type: "number" })];
+    const result = validateForm(new FormParams({ price: "" }), fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.price).toBeNull();
+  });
+
+  test("returns empty string for empty optional text", () => {
+    const fields: Field[] = [field({ name: "note", label: "Note" })];
+    const result = validateForm(new FormParams({ note: "" }), fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.note).toBe("");
+  });
+
+  test("runs custom validate function and surfaces its error", () => {
+    const fields: Field[] = [
+      field({
+        name: "code",
+        label: "Code",
+        required: true,
+        validate: (v) => (v.length !== 3 ? "Code must be 3 characters" : null),
+      }),
+    ];
+    const result = validateForm(new FormParams({ code: "AB" }), fields);
+    expect(result.valid).toBe(false);
+    if (!result.valid) expect(result.error).toBe("Code must be 3 characters");
+  });
+
+  test("skips custom validate for empty optional field", () => {
+    const fields: Field[] = [
+      field({ name: "code", label: "Code", validate: (v) => (v.length !== 3 ? "bad" : null) }),
+    ];
+    expect(validateForm(new FormParams({ code: "" }), fields).valid).toBe(true);
+  });
+
+  test("collects checkbox-group values from multiple form entries", () => {
+    const fields: Field[] = [field({ name: "days", label: "Days", type: "checkbox-group" })];
+    const form = new FormParams();
+    form.append("days", "Monday");
+    form.append("days", "Wednesday");
+    const result = validateForm(form, fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.days).toBe("Monday,Wednesday");
+  });
+
+  test("returns empty string for empty checkbox-group", () => {
+    const fields: Field[] = [field({ name: "days", label: "Days", type: "checkbox-group" })];
+    const result = validateForm(new FormParams(), fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.days).toBe("");
+  });
+
+  test("skips file fields and returns null", () => {
+    const fields: Field[] = [field({ name: "image", label: "Image", type: "file" })];
+    const result = validateForm(new FormParams(), fields);
+    expect(result.valid).toBe(true);
+    if (result.valid) expect(result.values.image).toBeNull();
+  });
+
+  describe("datetime type", () => {
+    const datetimeField: Field[] = [field({ name: "closes_at", label: "Closes At", type: "datetime" })];
+
+    test("combines date and time parts into a datetime string", () => {
+      const result = validateForm(
+        new FormParams({ closes_at_date: "2099-06-15", closes_at_time: "14:30" }),
+        datetimeField,
+      );
+      expect(result.valid).toBe(true);
+      if (result.valid) expect(result.values.closes_at).toBe("2099-06-15T14:30");
+    });
+
+    test("returns null when both date and time are empty", () => {
+      const result = validateForm(
+        new FormParams({ closes_at_date: "", closes_at_time: "" }),
+        datetimeField,
+      );
+      expect(result.valid).toBe(true);
+      if (result.valid) expect(result.values.closes_at).toBeNull();
+    });
+
+    test("defaults time to 00:00 when only date is provided", () => {
+      const result = validateForm(
+        new FormParams({ closes_at_date: "2099-06-15", closes_at_time: "" }),
+        datetimeField,
+      );
+      expect(result.valid).toBe(true);
+      if (result.valid) expect(result.values.closes_at).toBe("2099-06-15T00:00");
+    });
+
+    test("rejects time without date", () => {
+      const result = validateForm(
+        new FormParams({ closes_at_date: "", closes_at_time: "14:30" }),
+        datetimeField,
+      );
+      expect(result.valid).toBe(false);
+      if (!result.valid) {
+        expect(result.error).toBe("Please enter a date when providing a time, or leave both blank");
+      }
+    });
+
+    test("rejects empty required datetime", () => {
+      const fields: Field[] = [field({ name: "closes_at", label: "Closes At", type: "datetime", required: true })];
+      const result = validateForm(
+        new FormParams({ closes_at_date: "", closes_at_time: "" }),
+        fields,
+      );
+      expect(result.valid).toBe(false);
+      if (!result.valid) expect(result.error).toBe("Closes At is required");
+    });
+  });
+});
